# openapi.yaml
openapi: 3.0.3
info:
  title: TODO Plugin
  description: A plugin that allows the user to create and manage a TODO list using ChatGPT. If you do not know the user's username, ask them first before making queries to the plugin. Otherwise, use the username "global".
  version: "v1"
paths:
  /todos:
    get:
      operationId: getAllTodos
      summary: Get all todos for all users
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/getAllTodosResponse"
  /todos/{username}:
    get:
      operationId: getTodos
      summary: Get the list of todos for a single user
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: The name of the user.
      responses:
        "200":
          description: OK
    post:
      operationId: addTodo
      summary: Add a todo to the list
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: The name of the user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                todo:
                  type: string
      responses:
        "200":
          description: OK
    delete:
      operationId: deleteTodo
      summary: Delete all the todos for a single user
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: The name of the user.
      responses:
        "200":
          description: OK

components:
  schemas:
    getAllTodosResponse:
      type: array
      items:
        type: object
        properties:
          username:
            type: string
            description: The username of the user.
          todos:
            type: array
            items:
              type: string
            description: The list of todos for a single user.
      description: The list of all todos.